refactor(whatsrust): extract incoming message conversion
